Split object inspector panel
Extracted the selected-object editor from ObjectGUIManager into a dedicated InspectorPanel/ObjectInspector pair. The object tree now handles hierarchy and prefab actions only, while the new inspector owns component editing, add/remove flows, script drops, and viewport highlight state.
