+++ source/libs/editor/ObjectGUIManager.h
@@ -2,71 +2,47 @@
 #define OBJECTGUIMANAGER_H
 
 #include <nlohmann/json_fwd.hpp>
-#include <array>
 #include <functional>
 #include <memory>
 #include <optional>
-#include <unordered_set>
 #include <uuid.h>
 
 class ObjectManager;
-class AssetRegistry;
 class Object;
-class Component;
-class ComponentEditor;
 class EditorSelection;
 
-// The editor's object tree ("Objects") + selected-object panel ("Selected Object"). It draws each
-// component of the selected object through the ComponentEditor and reports edits back:
-//   - a component VALUE change -> EditCallback(objectUUID, component)
-//   - a STRUCTURAL change (add/remove object or component) -> SceneEditCallback(<built edit json>)
+// The editor's object tree ("Objects" panel): the hierarchy, its per-row context menu, "Save as Prefab",
+// and the delete-confirmation modal. It reports structural changes back:
+//   - a STRUCTURAL change (add/remove/reparent/duplicate object) -> SceneEditCallback(<built edit json>)
 //   - a new prefab asset ("Save as Prefab") -> AddAssetCallback(<built addAsset json>)
-// The EditorApp turns all three into network messages for the authoritative server.
+// The EditorApp turns both into network messages for the authoritative server. The selected object's
+// body is inspected by the separate InspectorPanel; both share the one EditorSelection slot.
 class ObjectGUIManager {
 public:
-  using EditCallback = std::function<void(const uuids::uuid& objectUUID, const std::shared_ptr<Component>& component)>;
   using SceneEditCallback = std::function<void(const nlohmann::json& edit)>;
   // Same blob (and same EditorApp handler) as AssetBrowserPanel::AddAssetCallback.
   using AddAssetCallback = std::function<void(const nlohmann::json& addAsset)>;
-
-  explicit ObjectGUIManager(std::shared_ptr<ComponentEditor> componentEditor);
-
-  void setEditCallback(EditCallback callback);
 
   void setSceneEditCallback(SceneEditCallback callback);
 
   void setAddAssetCallback(AddAssetCallback callback);
 
-  void setAssetRegistry(const AssetRegistry* registry);
-
-  // The shared selection slot owned by EditorApp: the tree click and delete flows read/write it, and
-  // the viewport highlight derives from it. Other writers (viewport picking, the asset browser) share
-  // the same instance, so selecting an object here and picking one there stay in sync.
+  // The shared selection slot owned by EditorApp: the tree click and delete flows read/write it. Other
+  // writers (viewport picking, the Inspector) share the same instance, so selecting an object here and
+  // picking one there stay in sync.
   void setSelection(std::shared_ptr<EditorSelection> selection);
 
-  // objectManager may be null (no scene loaded yet): the windows are still drawn, just empty, so they
-  // stay present/dockable instead of popping in and out.
+  // objectManager may be null (no scene loaded yet): the window is still drawn, just empty, so it stays
+  // present/dockable instead of popping in and out.
   void displayGui(const ObjectManager* objectManager);
 
-  // When false (the connected server isn't in edit mode), the panels still render so the scene can be
-  // viewed/inspected, but the add/remove/reparent/edit affordances are disabled.
+  // When false (the connected server isn't in edit mode), the tree still renders so the scene can be
+  // viewed/inspected, but the add/remove/reparent affordances are disabled.
   void setEditable(bool editable);
 
-  // The object to highlight in the viewport: the selection when "Highlight Object" is on, else nullopt.
-  [[nodiscard]] std::optional<uuids::uuid> getHighlightUUID() const;
-
 private:
-  std::shared_ptr<ComponentEditor> m_componentEditor;
-
-  EditCallback m_editCallback;
   SceneEditCallback m_sceneEditCallback;
   AddAssetCallback m_addAssetCallback;
-
-  const AssetRegistry* m_assetRegistry = nullptr;
-
-  // Buffer for in-place name editing. Refreshed whenever the selected object changes.
-  std::array<char, 256> m_nameEditBuffer{};
-  std::optional<uuids::uuid> m_nameEditObjectUUID;
 
   // The editor-wide selection, owned by EditorApp and shared with the other panels. The Object kind is
   // what this manager cares about; asset selections are inspected by other panels.
@@ -76,34 +52,18 @@ private:
   // While set, the "Delete Object?" modal is shown; confirming sends a removeObject scene edit.
   std::optional<uuids::uuid> m_objectPendingDeletion;
 
-  bool m_highlightSelectedObject = true;
-
   // False when the connected server is read-only (not in edit mode); gates the mutating UI.
   bool m_editable = true;
 
-  // Components whose deletion we've already sent (markedAsDeleted persists until the next snapshot
-  // rebuilds the object), so we don't re-send the same removeComponent every frame.
-  std::unordered_set<const Component*> m_pendingRemovals;
-
-  bool m_showComponentSelector = false;
-
   void displayObjectTree(const std::shared_ptr<Object>& object);
-
-  void displaySelectedObject(const ObjectManager* objectManager);
 
   // The "Delete Object?" confirmation modal for m_objectPendingDeletion. Confirming (Yes / Enter) sends
   // a removeObject scene edit; cancelling (No / Escape), or the object vanishing, clears the prompt.
   void displayDeleteConfirmationModal(const ObjectManager* objectManager);
 
-  void displayAddComponent(const std::shared_ptr<Object>& object);
-
   // Register the object's serialized blob as a Prefab asset (body carried inline, no file on disk).
   // Re-saving under the same name updates that prefab's body in place, keeping its uuid.
   void saveAsPrefab(const std::shared_ptr<Object>& object) const;
-
-  void displayScriptDragDropArea(float dropZoneStartY, const std::shared_ptr<Object>& object) const;
-
-  void displayComponent(const uuids::uuid& objectUUID, const std::shared_ptr<Component>& component);
 };
